pyproject.toml
[tool.poetry]
name = "dic"
version = "0.1.0"
description = ""
authors = ["Hordunlarmy <Hordunlarmy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["uvicorn"], version = "^0.115.6"}
python-decouple = "^3.8"
uvicorn = "^0.32.1"
inquirerpy = "^0.3.4"
psycopg2 = "^2.9.10"
python-multipart = "^0.0.19"
cloudinary = "^1.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
